# shared/api-contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "TTBS Customer API"
  description: "Customer-facing API contract for the train ticket booking system."
  version: "1.0"
servers:
- url: "/api"
  description: "Versioned customer API root"
security:
- bearerAuth: []
...
paths:
  /api/{version}/bookings:
    post:
      tags:
      - "Bookings"
      summary: "Create a booking"
      operationId: "createBooking"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
        required: true
      responses:
        "201":
          description: "Booking created"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: "Invalid booking payload"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Scheduled trip or customer not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "409":
          description: "Booking cannot be created because seats are unavailable or\
            \ already held"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
  /api/{version}/bookings/{id}/cancel:
    post:
      tags:
      - "Bookings"
      summary: "Cancel an existing booking"
      operationId: "cancelBooking"
      parameters:
      - name: "id"
        in: "path"
        description: "Booking identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Booking cancelled"
        "403":
          description: "Customer cannot cancel this booking"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Booking not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "409":
          description: "Booking cannot transition to cancelled state"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
...
  /api/{version}/users/{userId}/bookings:
    get:
      tags:
      - "Bookings"
      summary: "List bookings for a customer"
      operationId: "getUserBookings"
      parameters:
      - name: "userId"
        in: "path"
        description: "Customer identifier that owns the bookings"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetUserBookingsRequest"
      responses:
        "200":
          description: "Paged booking history"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseUserBookingResponse"
        "400":
          description: "Invalid pagination parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "403":
          description: "Customer cannot access another customer's bookings"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
...
  /api/{version}/bookings/{id}:
    get:
      tags:
      - "Bookings"
      summary: "Get a booking by id"
      operationId: "getBooking"
      parameters:
      - name: "id"
        in: "path"
        description: "Booking identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetBookingDetailRequest"
      responses:
        "200":
          description: "Booking detail"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetailResponse"
        "403":
          description: "Customer cannot access this booking"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Booking not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
    FailData:
      description: "Shared payload used inside JSend fail responses."
      properties:
        message:
          type: "string"
          description: "Human-readable summary of the failure."
        code:
          $ref: "#/components/schemas/ErrorCode"
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/Violation"
    ErrorCode:
      type: "string"
      description: "Machine-readable error code carried by shared fail responses."
      enum:
      - "VALIDATION_ERROR"
      - "CURSOR_INVALID"
      - "ACCESS_DENIED"
      - "USER_EMAIL_ALREADY_EXISTS"
      - "USER_INVALID_CREDENTIALS"
      - "USER_INVALID_REFRESH_TOKEN"
      - "USER_NOT_FOUND"
      - "USER_CANNOT_BULK_DELETE_SELF"
      - "USER_HAS_ACTIVE_BOOKINGS"
      - "BOOKING_NOT_FOUND"
      - "BOOKING_CANNOT_CONFIRM"
      - "BOOKING_ALREADY_CANCELLED"
      - "TRAIN_NUMBER_ALREADY_EXISTS"
      - "TRAIN_NOT_FOUND"
      - "TRAIN_IN_USE"
      - "SEAT_NOT_FOUND"
      - "SEAT_NUMBER_ALREADY_EXISTS"
      - "SEAT_NOT_AVAILABLE"
      - "SEAT_IN_USE"
      - "SCHEDULED_TRIP_NOT_FOUND"
      - "COACH_NOT_FOUND"
      - "COACH_CAR_NUMBER_ALREADY_EXISTS"
      - "COACH_TRAIN_NOT_FOUND"
      - "COACH_IN_USE"
      - "COACH_CAR_NUMBERS_ALREADY_EXIST"
      - "COACH_DUPLICATE_CAR_NUMBERS_IN_REQUEST"
      - "SEAT_NUMBERS_ALREADY_EXIST"
      - "SEAT_DUPLICATE_SEAT_NUMBERS_IN_REQUEST"
      - "ROUTE_TEMPLATE_NOT_FOUND"
      - "STATION_NOT_FOUND"
      - "STATION_CODE_ALREADY_EXISTS"
      - "STATION_IN_USE"
      - "PAYMENT_NOT_FOUND"
      - "PAYMENT_ALREADY_PROCESSED"
      - "PAYMENT_REFUND_FAILED"
    Violation:
      description: "Field-level validation issue inside a fail response."
      properties:
        field:
          type: "string"
          description: "Request field that failed validation."
        message:
          type: "string"
          description: "Human-readable validation message."
        code:
          $ref: "#/components/schemas/ViolationCode"
    ViolationCode:
      type: "string"
      description: "Machine-readable validation category attached to a field violation."
      enum:
      - "REQUIRED"
      - "INVALID_FORMAT"
      - "TOO_SHORT"
      - "TOO_LONG"
      - "INVALID_VALUE"
    JsendFailResponse:
      type: "object"
      description: "Shared JSend fail envelope used for validation and domain errors."
      properties:
        status:
          type: "string"
          enum:
          - "fail"
        data:
          $ref: "#/components/schemas/FailData"
    JsendErrorResponse:
      type: "object"
      description: "Shared JSend error envelope used for unexpected technical failures."
      properties:
        status:
          type: "string"
          enum:
          - "error"
        message:
          type: "string"
          description: "Human-readable technical failure message."
    CreateBookingRequest:
      type: "object"
      description: "Booking creation payload."
      properties:
        scheduledTripId:
          type: "string"
          format: "uuid"
          description: "Scheduled trip to book."
        seatIds:
          type: "array"
          items:
            type: "string"
            format: "uuid"
            description: "Seat identifier."
          minItems: 1
        idempotencyKey:
          type: "string"
          description: "Client-generated idempotency key used to safely retry booking\
            \ creation."
          example: "booking-create-request-001"
          minLength: 1
          writeOnly: true
      required:
      - "idempotencyKey"
      - "scheduledTripId"
      - "seatIds"
    JsendResponseObject:
      type: "object"
      description: "Runtime JSend response envelope used by the API implementation."
      properties:
        status:
          type: "string"
          description: "JSend status value."
          enum:
          - "success"
          - "fail"
          - "error"
        data:
          description: "Payload returned for the response status."
        message:
          type: "string"
          description: "Human-readable technical error message for error responses."
...
    GetUserBookingsRequest:
      type: "object"
      description: "Offset-based pagination query for a customer's booking history."
      properties:
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Number of bookings per page."
          example: 20
          maximum: 100
          minimum: 1
...
    GetBookingDetailRequest:
      description: "No additional query parameters are accepted when fetching a booking\
        \ detail."
...
    BookingResponse:
      description: "Booking resource returned after successful booking creation."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Booking identifier."
          readOnly: true
        userId:
          type: "string"
          format: "uuid"
          description: "Customer identifier that owns the booking."
        scheduledTripId:
          type: "string"
          format: "uuid"
          description: "Scheduled trip identifier."
        userInfo:
          $ref: "#/components/schemas/PassengerInfoResponse"
        totalPrice:
          type: "integer"
          format: "int64"
          description: "Booking total in minor currency units."
          example: 650000
        currency:
          type: "string"
          description: "ISO-like currency code."
          example: "VND"
        status:
          type: "string"
          description: "Booking lifecycle status."
          enum:
          - "HELD"
          - "CONFIRMED"
          - "CANCELLED"
        paymentDeadline:
          type: "string"
          format: "date-time"
          description: "Deadline for completing payment before the booking expires."
        createdAt:
          type: "string"
          format: "date-time"
          description: "Booking creation timestamp."
          readOnly: true
    PassengerInfoResponse:
      description: "Passenger identity and contact information captured for a booking."
      properties:
        fullName:
          type: "string"
          description: "Passenger full name."
          example: "Nguyen Phuong"
        email:
          type: "string"
          description: "Passenger email address."
          example: "customer@example.com"
        phone:
          type: "string"
          description: "Passenger phone number."
          example: "+84901234567"
        dateOfBirth:
          type: "string"
          format: "date"
          description: "Passenger date of birth."
        gender:
          type: "string"
          description: "Passenger self-declared gender label."
          example: "female"
        idDocumentNumber:
          type: "string"
          description: "Passenger government-issued identity document number."
          example: "redacted-id-document"
        addressLine:
          type: "string"
          description: "Passenger address line."
          example: "redacted-address"
...
    PageResponseUserBookingResponse:
      description: "Offset-based pagination wrapper."
      properties:
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/UserBookingResponse"
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Requested page size."
          example: 20
          minimum: 1
        hasNext:
          type: "boolean"
          description: "Whether a subsequent page exists."
        hasPrevious:
          type: "boolean"
          description: "Whether a previous page exists."
        total:
          type: "integer"
          format: "int64"
          description: "Total number of matching items across all pages."
          example: 125
          minimum: 0
    UserBookingResponse:
      description: "Compact booking summary for booking-history listings."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Booking identifier."
          readOnly: true
        userId:
          type: "string"
          format: "uuid"
          description: "Customer identifier that owns the booking."
        scheduledTripId:
          type: "string"
          format: "uuid"
          description: "Scheduled trip identifier."
        totalPrice:
          type: "integer"
          format: "int64"
          description: "Booking total in minor currency units."
          example: 650000
        currency:
          type: "string"
          description: "ISO-like currency code."
          example: "VND"
        status:
          type: "string"
          description: "Booking lifecycle status."
          enum:
          - "HELD"
          - "CONFIRMED"
          - "CANCELLED"
        paymentDeadline:
          type: "string"
          format: "date-time"
          description: "Deadline for completing payment before the booking expires."
        createdAt:
          type: "string"
          format: "date-time"
          description: "Booking creation timestamp."
          readOnly: true
...
    Route:
      description: "Route summary embedded inside a search result."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Route template identifier."
        basePrice:
          type: "integer"
          format: "int64"
          description: "Base fare in minor currency units."
          example: 650000
        currency:
          type: "string"
          description: "ISO-like currency code."
          example: "VND"
        origin:
          $ref: "#/components/schemas/Station"
        destination:
          $ref: "#/components/schemas/Station"
...
    Station:
      description: "Station summary embedded inside a route."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Station identifier."
        code:
          type: "string"
          description: "Station code."
          example: "HNO"
        name:
          type: "string"
          description: "Station name."
          example: "Ha Noi"
        city:
          type: "string"
          description: "Station city."
          example: "Ha Noi"
    Train:
      description: "Train summary embedded inside a search result."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Train identifier."
        trainNumber:
          type: "string"
          description: "Train number."
          example: "SE1"
        name:
          type: "string"
          description: "Train display name."
          example: "Reunification Express"
        totalSeats:
          type: "integer"
          format: "int32"
          description: "Total seats available on the train."
          example: 320
...
    Seat:
      description: "Seat state inside a coach seat map."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Seat identifier."
        seatNumber:
          type: "string"
          description: "Seat label shown to customers."
          example: "12A"
        status:
          type: "string"
          description: "Current seat availability status."
          enum:
          - "AVAILABLE"
          - "HELD"
          - "BOOKED"
          - "CANCELLED"
...
    BookingDetailResponse:
      description: "Detailed booking resource with trip, seat, and payment information."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Booking identifier."
          readOnly: true
        userId:
          type: "string"
          format: "uuid"
          description: "Customer identifier that owns the booking."
        scheduledTripId:
          type: "string"
          format: "uuid"
          description: "Scheduled trip identifier."
        passengerInfo:
          $ref: "#/components/schemas/PassengerInfoResponse"
        totalPrice:
          type: "integer"
          format: "int64"
          description: "Booking total in minor currency units."
          example: 650000
        currency:
          type: "string"
          description: "ISO-like currency code."
          example: "VND"
        status:
          type: "string"
          description: "Booking lifecycle status."
          enum:
          - "HELD"
          - "CONFIRMED"
          - "CANCELLED"
        paymentDeadline:
          type: "string"
          format: "date-time"
          description: "Deadline for completing payment before the booking expires."
        createdAt:
          type: "string"
          format: "date-time"
          description: "Booking creation timestamp."
          readOnly: true
        trip:
          $ref: "#/components/schemas/Trip"
        payment:
          $ref: "#/components/schemas/PaymentDetailResponse"
        seats:
          type: "array"
          items:
            $ref: "#/components/schemas/Seat"
    PaymentDetailResponse:
      description: "Payment snapshot associated with a booking."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Payment identifier."
          readOnly: true
        status:
          type: "string"
          description: "Payment lifecycle status."
          enum:
          - "PENDING"
          - "PAID"
          - "CANCELLED"
          - "FAILED"
          - "REFUNDED"
        checkoutUrl:
          type: "string"
          format: "uri"
          description: "Hosted checkout URL when further payment action is required."
        amount:
          type: "integer"
          format: "int64"
          description: "Payment amount in minor currency units."
          example: 650000
        currency:
          type: "string"
          description: "ISO-like currency code."
          example: "VND"
        stripePaymentIntentId:
          type: "string"
          description: "Stripe payment intent identifier used for reconciliation."
        createdAt:
          type: "string"
          format: "date-time"
          description: "Payment creation timestamp."
          readOnly: true
    Trip:
      description: "Trip summary embedded inside a booking detail."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Scheduled trip identifier."
        routeTemplateId:
          type: "string"
          format: "uuid"
          description: "Route template identifier."
        trainId:
          type: "string"
          format: "uuid"
          description: "Train identifier."
        departureTime:
          type: "string"
          format: "date-time"
          description: "Scheduled departure timestamp."
        arrivalTime:
          type: "string"
          format: "date-time"
          description: "Scheduled arrival timestamp."
        status:
          type: "string"
          description: "Scheduled trip status."
        createdAt:
          type: "string"
          format: "date-time"
          description: "Scheduled trip creation timestamp."
        train:
          $ref: "#/components/schemas/Train"
        route:
          $ref: "#/components/schemas/Route"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"